test: move controller fixtures to common header
